LayoutTests/imported/w3c/web-platform-tests/navigation-api/commit-behavior/after-transition-uncancelable-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: InvalidStateError: Cannot go back


FAIL { commit: 'after-transition' } for an uncancelable traverse navigation promise_test: Unhandled rejection with value: object "InvalidStateError: Cannot go back"
